fix(misc): fixed old sidecar imports and renamed dockerignore
