+++ cmd/sidecars/lbsidecar/.dockerignore
